+++ stl/lib/stl.dart
@@ -47,3 +47,4 @@ export 'src/primitives/uint8.dart';
 export 'src/primitives/uint16.dart';
 export 'src/primitives/uint32.dart';
 export 'src/primitives/uint64.dart';
+export 'src/math/constant/constant.dart';